+++ common/src/main/java/com/fej1fun/potentials/components/FluidAmountMapDataComponent.java
@@ -3,7 +3,6 @@ package com.fej1fun.potentials.components;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import dev.architectury.fluid.FluidStack;
-import net.minecraft.core.Holder;
 import net.minecraft.core.NonNullList;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
@@ -16,23 +15,21 @@ import net.minecraft.world.level.material.Fluids;
 import java.io.Serializable;
 import java.util.*;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class FluidAmountMapDataComponent implements Serializable {
 
-    @SuppressWarnings("all")
     public static final Codec<FluidAmountMapDataComponent> CODEC = RecordCodecBuilder.create(instance -> instance.group(
-            BuiltInRegistries.FLUID.holderByNameCodec().listOf().fieldOf("fluids")
+            BuiltInRegistries.FLUID.byNameCodec().listOf().fieldOf("fluids")
                     .forGetter(component ->
-                        component.fluids.stream().map(fluid -> (Holder<Fluid>) fluid.builtInRegistryHolder()).toList()),
+                        component.fluids.stream().toList()),
             Codec.LONG.listOf().fieldOf("amounts").forGetter(component -> component.amounts)
-    ).apply(instance, FluidAmountMapDataComponent::create));
-    @SuppressWarnings("all")
+    ).apply(instance, FluidAmountMapDataComponent::new));
+
     public static final StreamCodec<RegistryFriendlyByteBuf, FluidAmountMapDataComponent> STREAM_CODEC = StreamCodec.composite(
-            ByteBufCodecs.holderRegistry(Registries.FLUID).apply(ByteBufCodecs.list()), component ->
-                    component.fluids.stream().map(fluid-> (Holder<Fluid>) fluid.builtInRegistryHolder()).toList(),
+            ByteBufCodecs.registry(Registries.FLUID).apply(ByteBufCodecs.list()), component ->
+                    component.fluids.stream().toList(),
             ByteBufCodecs.VAR_LONG.apply(ByteBufCodecs.list()), component -> component.amounts,
-            FluidAmountMapDataComponent::create
+            FluidAmountMapDataComponent::new
     );
 
     protected final List<Fluid> fluids;
@@ -52,13 +49,6 @@ public class FluidAmountMapDataComponent implements Serializable {
         this.amounts = Arrays.asList(amounts.toArray(new Long[0]));
     }
 
-    public static FluidAmountMapDataComponent create(List<Holder<Fluid>> holders, List<Long> amounts) {
-        if (holders.isEmpty()) throw new IllegalArgumentException("List<Holder<Fluid>> Cannot be empty");
-        if (amounts.isEmpty()) throw new IllegalArgumentException("List<Long> Cannot be empty");
-        if (holders.size() != amounts.size()) throw new IllegalArgumentException("List<Holder<Fluid>> must be the same length as List<Long>");
-
-        return new FluidAmountMapDataComponent(holders.stream().map(Holder::value).toList(), amounts);
-    }
     public static FluidAmountMapDataComponent create(final List<FluidStack> fluidStacks) {
         if (fluidStacks.isEmpty()) throw new IllegalArgumentException("List<FluidStack> Cannot be empty");
 
@@ -93,18 +83,18 @@ public class FluidAmountMapDataComponent implements Serializable {
 
     public FluidAmountMapDataComponent setFluid(int i, Fluid fluid) {
         this.fluids.set(i, fluid);
-        return this;
+        return new FluidAmountMapDataComponent(this.fluids, this.amounts);
     }
 
     public FluidAmountMapDataComponent setAmount(int i, Long amount) {
         this.amounts.set(i, amount);
-        return this;
+        return new FluidAmountMapDataComponent(this.fluids, this.amounts);
     }
 
     public FluidAmountMapDataComponent setFluidStack(int i, FluidStack stack) {
         setFluid(i, stack.getFluid());
         setAmount(i, stack.getAmount());
-        return this;
+        return new FluidAmountMapDataComponent(this.fluids, this.amounts);
     }
 
     @Override
@@ -113,14 +103,6 @@ public class FluidAmountMapDataComponent implements Serializable {
             return component.fluids.equals(this.fluids) && component.amounts.equals(this.amounts);
         return super.equals(obj);
     }
-
-    private static <T> List<T> getListFromStream(Stream<T> s) {
-        ArrayList<T> list = new ArrayList<>(0);
-        s.forEach(list::add);
-        list.trimToSize();
-        return list;
-    }
-
 
     @Override
     public int hashCode() {