+++ neoforge/src/main/java/com/fej1fun/potentials/neoforge/energy/NeoForgeEnergyStorage.java
@@ -1,35 +1,65 @@
 package com.fej1fun.potentials.neoforge.energy;
 
-import com.fej1fun.potentials.energy.UniversalEnergyStorage;
+import com.fej1fun.potentials.energy.BaseEnergyStorage;
 import net.neoforged.neoforge.transfer.energy.EnergyHandler;
-import net.neoforged.neoforge.transfer.transaction.Transaction;
+import net.neoforged.neoforge.transfer.transaction.SnapshotJournal;
 import net.neoforged.neoforge.transfer.transaction.TransactionContext;
 import org.jetbrains.annotations.NotNull;
 
 public class NeoForgeEnergyStorage implements EnergyHandler {
-    final UniversalEnergyStorage universalEnergyStorage;
+    private final SnapshotJournal<Integer> snapshotJournal = new SnapshotJournal<>() {
+        @Override
+        protected Integer createSnapshot() {
+            return baseEnergyStorage.getEnergy();
+        }
 
-    public NeoForgeEnergyStorage(@NotNull UniversalEnergyStorage universalEnergyStorage) {
-       this.universalEnergyStorage = universalEnergyStorage;
+        @Override
+        protected void revertToSnapshot(Integer snapshot) {
+            baseEnergyStorage.setEnergyStored(snapshot);
+        }
+    };
+
+    private final BaseEnergyStorage baseEnergyStorage;
+
+    public NeoForgeEnergyStorage(@NotNull BaseEnergyStorage baseEnergyStorage) {
+        this.baseEnergyStorage = baseEnergyStorage;
     }
 
     @Override
     public long getAmountAsLong() {
-        return universalEnergyStorage.getEnergy();
+        return baseEnergyStorage.getEnergy();
     }
 
     @Override
     public long getCapacityAsLong() {
-        return universalEnergyStorage.getMaxEnergy();
+        return baseEnergyStorage.getMaxEnergy();
     }
 
     @Override
     public int insert(int toReceive, TransactionContext transactionContext) {
-        return universalEnergyStorage.insert(toReceive, false);
+        return insert(toReceive, transactionContext, false);
     }
 
     @Override
-    public int extract(int toReceive, TransactionContext transactionContext) {
-       return  universalEnergyStorage.extract(toReceive, false);
+    public int extract(int toExtract, TransactionContext transactionContext) {
+        return extract(toExtract, transactionContext, false);
+    }
+
+    public int insert(int toReceive, TransactionContext transactionContext, boolean simulate) {
+        int amount = baseEnergyStorage.insert(toReceive, simulate);
+        if (amount > 0 && !simulate) {
+            snapshotJournal.updateSnapshots(transactionContext);
+        }
+
+        return amount;
+    }
+
+    public int extract(int toExtract, TransactionContext transactionContext, boolean simulate) {
+        int amount = baseEnergyStorage.extract(toExtract, simulate);
+        if (amount > 0 && !simulate) {
+            snapshotJournal.updateSnapshots(transactionContext);
+        }
+
+        return amount;
     }
 }